Create Surrogate Learner

Creates the default mlrMBO surrogate learners as an mlr3::Learner.

This imitates the behaviour of mlrCPO when no learner argument is given to mbo() / initSMBO().

makeMlr3Surrogate(
  is.numeric = TRUE,
  is.noisy = TRUE,
  has.dependencies = !is.numeric
)

Arguments

is.numeric
(logical(1))
Whether only numeric parameters are present. If so, a LearnerRegrKM (DiceKriging package) is constructed. Otherwise a LearnerRegrRanger (random forest from the ranger package) is constructed. Default is TRUE.

is.noisy
(logical(1))
Whether to use nugget estimation. Only considered when is.numeric is TRUE. Default is TRUE.

has.dependencies
(logical(1))
Whether to anticipate missing values in the surrogate model design. This adds out-of-range imputation to the model. If more elaborate imputation is desired, it may be desirable to set this to FALSE and instead perform custom imputation using mlr3pipelines. Default is !numeric.

Examples

# DiceKriging Learner:
makeMlr3Surrogate()
#> <LearnerRegrKM:regr.km>
#> * Model: -
#> * Parameters: covtype=matern3_2, optim.method=gen, jitter=1e-12,
#>   nugget.estim=TRUE
#> * Packages: mlr3, DiceKriging
#> * Predict Type: se
#> * Feature types: logical, integer, numeric
#> * Properties: -

# mlr3pipelines Graph: imputation %>>% 'ranger' (randomForest):
makeMlr3Surrogate(is.numeric = FALSE)
#> <GraphLearner:imputeoor.regr.ranger>
#> * Model: -
#> * Parameters: imputeoor.min=TRUE, imputeoor.offset=1,
#>   imputeoor.multiplier=1, regr.ranger.se.method=jack,
#>   regr.ranger.keep.inbag=TRUE
#> * Packages: mlr3, mlr3pipelines
#> * Predict Type: se
#> * Feature types: logical, integer, numeric, character, factor, ordered,
#>   POSIXct
#> * Properties: featureless, hotstart_backward, hotstart_forward,
#>   importance, loglik, missings, oob_error, selected_features, weights

# just the 'ranger' Learner:
makeMlr3Surrogate(is.numeric = FALSE, has.dependencies = FALSE)
#> <LearnerRegrRanger:regr.ranger>
#> * Model: -
#> * Parameters: se.method=jack, keep.inbag=TRUE
#> * Packages: mlr3, ranger
#> * Predict Type: se
#> * Feature types: logical, integer, numeric, character, factor, ordered
#> * Properties: importance, oob_error, weights